jdfbot-master/dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file dev-requirements.txt dev-requirements.in
#
alembic==1.0.5
appdirs==1.4.3            # via black
argh==0.26.2              # via watchdog
aspy.yaml==1.1.1          # via pre-commit
astroid==2.1.0            # via pylint
attrs==18.2.0             # via black
autoflake==1.2
autopep8==1.4.3
backcall==0.1.0           # via ipython
black==18.9b0
cached-property==1.5.1    # via pre-commit
certifi==2018.11.29
cfgv==1.1.0               # via pre-commit
chardet==3.0.4
click==7.0
decorator==4.3.0          # via ipython, traitlets
dparse==0.4.1             # via safety
fbmq==2.5.0
flake8==3.6.0
flask-migrate==2.3.1
flask-sqlalchemy==2.3.2
flask==1.0.2
identify==1.1.7           # via pre-commit
idna==2.8
importlib-metadata==0.7   # via pre-commit
ipdb==0.11
ipython-genutils==0.2.0   # via traitlets
ipython==7.2.0
isort==4.3.4
itsdangerous==1.1.0
jedi==0.13.2              # via ipython
jinja2==2.10
lazy-object-proxy==1.3.1  # via astroid
logzero==1.5.0
mako==1.0.7
markupsafe==1.1.0
mccabe==0.6.1             # via flake8, pylint
mypy-extensions==0.4.1    # via mypy
mypy==0.650
nodeenv==1.3.3            # via pre-commit
packaging==18.0           # via dparse, safety
parso==0.3.1              # via jedi
pathtools==0.1.2          # via watchdog
pep257==0.7.0
pexpect==4.6.0            # via ipython
pickleshare==0.7.5        # via ipython
pip-tools==3.2.0
pre-commit==1.13.0
prompt-toolkit==2.0.7     # via ipython
psycopg2-binary==2.7.6.1
ptvsd==4.2.0
ptyprocess==0.6.0         # via pexpect
pycodestyle==2.4.0        # via autopep8, flake8
pyflakes==2.0.0           # via autoflake, flake8
pygments==2.3.1           # via ipython
pylint==2.2.2
pyparsing==2.3.0          # via packaging
python-dateutil==2.7.5
python-dotenv==0.10.1
python-editor==1.0.3
pyyaml==3.13              # via aspy.yaml, dparse, pre-commit, watchdog
requests==2.21.0
rope==0.11.0
safety==1.8.4
six==1.12.0
sqlalchemy==1.2.15
toml==0.10.0              # via black, pre-commit
traitlets==4.3.2          # via ipython
typed-ast==1.1.1          # via mypy
urllib3==1.24.1
virtualenv==16.1.0        # via pre-commit
watchdog==0.9.0
wcwidth==0.1.7            # via prompt-toolkit
werkzeug==0.14.1
wrapt==1.10.11            # via astroid
yapf==0.25.0
